move context.py outside
