+++ mars/oscar/backends/ray/backend.py
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .driver import RayActorDriver
-from ..mars.context import MarsActorContext
 from ...backend import BaseActorBackend, register_backend
+from ..context import MarsActorContext
+from .driver import RayActorDriver
 
 __all__ = ['RayActorBackend']
